requirements.txt
discord.py>=2.6.4
python-dotenv>=1.1.1
httpx>=0.27.0
aiohttp>=3.13.1
